% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFieldNameMapping.R
\name{getFieldNameMapping}
\alias{getFieldNameMapping}
\title{The function getFieldNameMapping is tabken from
https://github.com/davidcurrie2001/MI_RDBES_ExchangeFiles in 6/11/2020,
with minor changes.}
\usage{
getFieldNameMapping(
  downloadFromGitHub = TRUE,
 
    gitHubDirectory = "https://api.github.com/repos/ices-tools-dev/RDBES/contents/Documents",
  fileLocation
)
}
\arguments{
\item{downloadFromGitHub}{TRUE if you want to download the latest data model
spreadsheets from GitHub}

\item{gitHubDirectory}{(Optional) Default value is
"https://api.github.com/repos/ices-tools-dev/RDBES/contents/Documents"}

\item{fileLocation}{The location you want to save and read the data model
spredsheet from}
}
\value{
list of dataframes mapping the variables of RDBES exchange format.
}
\description{
getFieldNameMapping Get a data frame that maps between database names and the
shorter R names
}
