% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRDBESFile.R
\name{importRDBESFile}
\alias{importRDBESFile}
\title{Read a RDBES import file and store it as a list with a dataframe by table.}
\usage{
importRDBESFile(file, path = getwd())
}
\arguments{
\item{file}{vector with the Accidentals filenames}

\item{path}{path of the files. The working directory by default.}
}
\value{
Return list of dataframes
}
\description{
Read the data as is, without any kind of validation. The purpouse of this
function is facilitate a simple validation by table.
}
